enums data added and pass the correct values to method for connect, disconnect, run and stop

+++ InvertekDrive-RTU-UI/InvertekDrive-RTU-UI/Services/ModbusService.cs
@@ -33,9 +33,9 @@ public static class ModbusService
 
     public static bool ConnectModbusMaster(string selectedPort,
         int baudRate,
-        int dataBits,
+        ushort dataBits,
         string parity,
-        int stopBits
+        ushort stopBits
     )
     {
         try
